Rename package aas-sm-to-py to aas-submodel-to-py
Expand the abbreviated 'sm' to 'submodel' for clarity across the
entire repository: directory, PyPI package name, importable Python
module (aas_submodel_to_py), CI workflow, and all documentation.

README improvements included in this change:

README.md: added PyPI badges, replaced vague "Features" section with
a packages table, added Quick Start section with install/usage
examples, and a repository structure diagram.

aas-submodel-to-py/README.md: mentioned .json/.xml support in the
intro, restructured Usage section with short-flag CLI examples and a
flag reference table, documented the previously undocumented
generate_from_obj_store method in the Python API section.

py-aas-submodels/README.md: rewritten with a proper intro,
installation instructions, a table of all 8 available submodels with
class names, a complete usage example using the real Nameplate
constructor signature, and a pointer to aas-submodel-to-py for
generating additional classes.
