+++ .github/workflows/publish.yml
@@ -32,8 +32,8 @@ jobs:
         with:
           packages-dir: py-aas-submodels/dist
 
-  publish-aas-sm-to-py:
-    name: Upload aas-sm-to-py to PyPI
+  publish-aas-submodel-to-py:
+    name: Upload aas-submodel-to-py to PyPI
     runs-on: ubuntu-latest
     environment: release
     permissions:
@@ -51,10 +51,10 @@ jobs:
           python -m pip install --upgrade pip
           pip install setuptools wheel build
       - name: Create source and wheel dist
-        working-directory: aas-sm-to-py
+        working-directory: aas-submodel-to-py
         run: |
           python -m build
       - name: Publish distribution to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          packages-dir: aas-sm-to-py/dist
+          packages-dir: aas-submodel-to-py/dist
