+++ aas-submodel-to-py/aas_submodel_to_py/__init__.py
