+++ aas-submodel-to-py/aas_submodel_to_py/code_templates/referable_with_se_class.pyi
