+++ aas-submodel-to-py/aas_submodel_to_py/code_templates/se_col_class.pyi
