+++ aas-submodel-to-py/aas_submodel_to_py/code_templates/se_list_class.pyi
