+++ aas-submodel-to-py/aas_submodel_to_py/code_templates/submodel_class.pyi
