+++ aas-submodel-to-py/aas_submodel_to_py/util.py
